Use integer instead of bigint where unneeded, Generate migrations

+++ core/model/filter_word.py
@@ -1,5 +1,5 @@
 from .import Base
-from sqlalchemy import Column, BigInteger, String, Boolean, Index
+from sqlalchemy import Column, BigInteger, String, Boolean, Index, Integer
 
 
 class FilterWord(Base):
@@ -8,6 +8,6 @@ class FilterWord(Base):
     phrase = Column(String, primary_key=True, index=True)
     should_notify = Column(Boolean, nullable=False)
     filter_without_removing = Column(Boolean, default=False)
-    bypass_level = Column(BigInteger, nullable=False)
+    bypass_level = Column(Integer, nullable=False)
     disable_extra_checks = Column(Boolean, default=False)
     is_piracy_phrase = Column(Boolean, default=False)
