+++ core/model/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, String, Boolean
+from sqlalchemy import Column, BigInteger, String, Boolean, Integer
 
 from . import Base
 
@@ -9,13 +9,13 @@ class User(Base):
     user_id = Column(BigInteger, primary_key=True, index=True)
     username = Column(String)
     is_clem = Column(Boolean, default=False)
-    xp = Column(BigInteger, default=-1)
-    level = Column(BigInteger, default=-1)
+    xp = Column(Integer, default=-1)
+    level = Column(Integer, default=-1)
     is_xp_frozen = Column(Boolean, default=False)
     was_warn_kicked = Column(Boolean, default=False)
     is_birthday_banned = Column(Boolean, default=False)
     is_raid_verified = Column(Boolean, default=False)
-    warn_points = Column(BigInteger, default=-1)
+    warn_points = Column(Integer, default=-1)
     timezone = Column(String)
     should_offline_report_ping = Column(Boolean, default=False)
